% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readhis.R
\name{readhis}
\alias{readhis}
\title{Reads Sobek his output files into Large Array's}
\usage{
readhis(filename)
}
\arguments{
\item{filename}{The folder and filename of the Sobek HIS output file}
}
\description{
Reads Sobek his output files into Large Array's
}
