% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2arr.R
\name{map2arr}
\alias{map2arr}
\title{Reads Sobek/Delwaq map output files into Array's}
\usage{
map2arr(filename, timestamp = T, begintime = "1900-01-01 00:00:00")
}
\description{
Reads Sobek/Delwaq map output files into Array's
}
